+++ backend/playdeep/schemas.py
@@ -1,0 +1,15 @@
+from ninja import Schema
+
+class UserSchema(Schema):
+    username: str
+    is_authenticated: bool
+    email: str = None
+
+class RegisterSchema(Schema):
+    username: str
+    email: str
+    password: str
+
+class LoginSchema(Schema):
+    username: str
+    password: str
