+++ backend/video/models.py
@@ -3,6 +3,8 @@ from django.db import models
 from django.urls import reverse
 from .validators import file_size
 from .storage_backend import AzureMediaStorage
+from django.contrib.auth import get_user_model
+
 # Create your models here.
 
 class Video(models.Model):
@@ -13,6 +15,11 @@ class Video(models.Model):
         max_length=250,
         unique_for_date='publish'
     )
+    user = models.ForeignKey(
+    get_user_model(),
+    on_delete=models.CASCADE,
+    related_name="videos"
+)
     def __str__(self):
         return self.caption
